backend/requirements.txt
# Основные зависимости
flask==2.2.3
python-dotenv==1.0.0
pydub==0.25.1

# OAuth2 зависимости
authlib==1.3.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Whisper и зависимости
openai-whisper==20231117
PyAudio==0.2.13
sounddevice==0.4.6
webrtcvad==2.0.10

# Для работы с данными
numpy==1.24.2

click==8.1.7
fastapi==0.104.1
uvicorn==0.24.0
jiwer==3.0.3
python-json-logger==2.0.7
pyinstaller==6.12.0
python-multipart==0.0.9

# Для улучшенной системы тегирования
nltk==3.8.1
pymorphy2==0.9.1
scikit-learn==1.3.2 